Rearrange vertx handler advice
